requirements.txt
Django==6.0
djangorestframework==3.16.1
django-cors-headers==4.9.0
openai==2.11.0
python-dotenv==1.2.1

